refactor: separate styles from components
