+++ src/components/box/index.tsx
@@ -2,12 +2,12 @@ import React, { PropsWithChildren } from 'react';
 import styled from 'styled-components';
 import { Styles, withStyles } from '@lapidist/styles';
 
+export * from './styles';
+
 export interface BoxProps extends PropsWithChildren<unknown> {
     readonly styles?: Styles;
     readonly as?: keyof JSX.IntrinsicElements;
 }
-
-export const boxStyles = (props: BoxProps): Styles => props.styles || {};
 
 export const Box: React.FC<BoxProps> = styled.div<BoxProps>`
     ${({ theme, styles }): string => withStyles(theme, styles)}
