+++ src/components/box/styles.ts
@@ -1,0 +1,4 @@
+import { Styles } from '@lapidist/styles';
+import { BoxProps } from './index';
+
+export const boxStyles = (props: BoxProps): Styles => props.styles || {};
