+++ src/components/button/index.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import deepMerge from 'lodash.merge';
 import { withTheme } from 'styled-components';
-import { ColorGroup, getProperty, Styles, Theme } from '@lapidist/styles';
+import { Theme } from '@lapidist/styles';
 import { Text } from '../text';
 import { BoxProps } from '../box';
+import { buttonStyles } from './styles';
+
+export * from './styles';
 
 export type ButtonPropType = BoxProps &
     React.ButtonHTMLAttributes<HTMLButtonElement> &
@@ -15,104 +18,6 @@ export interface ButtonProps {
     readonly small?: boolean;
     readonly ghost?: boolean;
 }
-
-interface ButtonVariantStyles {
-    borderColor: ColorGroup | string;
-    backgroundColor: ColorGroup | string;
-    paddingY: string;
-    fontSize: string;
-    textColor: ColorGroup | string;
-    hoverBackgroundColor: ColorGroup | string;
-    hoverTextColor: ColorGroup | string;
-    disabledHoverBackgroundColor: ColorGroup | string;
-    disabledTextColor: ColorGroup | string;
-    disabledBackgroundColor: ColorGroup | string;
-}
-
-const buttonSizing = (small?: boolean) => ({
-    paddingY: small ? '1' : '2',
-    fontSize: small ? '2' : '3'
-});
-
-const darkGrey: ColorGroup = { group: 'grey', shade: 'dark' };
-
-const lightBase: ColorGroup = { group: 'base', shade: 'light' };
-
-const lightestGrey: ColorGroup = { group: 'grey', shade: 'lightest' };
-
-const buttonColors = (dark: string, base: string, ghost?: boolean) => ({
-    borderColor: dark,
-    hoverBackgroundColor: dark,
-    hoverTextColor: lightBase,
-    backgroundColor: ghost ? 'transparent' : base,
-    textColor: ghost ? darkGrey : lightBase,
-    disabledHoverBackgroundColor: ghost ? lightestGrey : dark,
-    disabledTextColor: ghost ? darkGrey : lightBase,
-    disabledBackgroundColor: ghost ? lightestGrey : dark
-});
-
-const buttonVariants = ({
-    kind,
-    theme,
-    small,
-    ghost
-}: ButtonProps): ButtonVariantStyles => {
-    const { dark, base } = getProperty<{
-        [K: string]: string;
-    }>(theme, 'colors', kind);
-
-    return {
-        ...buttonSizing(small),
-        ...buttonColors(dark, base, ghost)
-    };
-};
-
-const buttonBaseStyles: Styles = {
-    borderWidth: '1',
-    textAlign: 'center',
-    paddingX: '4',
-    borderRadius: '2',
-    borderStyle: 'solid'
-};
-
-const buttonVariantStyles = ({
-    borderColor,
-    backgroundColor,
-    textColor,
-    paddingY,
-    fontSize,
-    hoverBackgroundColor,
-    hoverTextColor,
-    disabledHoverBackgroundColor,
-    disabledTextColor,
-    disabledBackgroundColor
-}: ButtonVariantStyles): Styles => ({
-    ...buttonBaseStyles,
-    borderColor,
-    backgroundColor,
-    textColor,
-    paddingY,
-    fontSize,
-    pseudo: {
-        ':hover': {
-            cursor: 'pointer',
-            backgroundColor: hoverBackgroundColor,
-            textColor: hoverTextColor
-        },
-        ':hover:disabled': {
-            cursor: 'not-allowed',
-            backgroundColor: disabledHoverBackgroundColor
-        },
-        ':disabled': {
-            opacity: '0.7',
-            textColor: disabledTextColor,
-            backgroundColor: disabledBackgroundColor
-        }
-    }
-});
-
-export const buttonStyles = (props: ButtonProps): Styles =>
-    buttonVariantStyles(buttonVariants(props));
 
 const BaseButton: React.FC<ButtonPropType & ButtonProps> = ({
     as = 'button',
