+++ src/components/flex/styles.ts
@@ -1,0 +1,5 @@
+import { Styles } from '@lapidist/styles';
+
+export const flexStyles = (): Styles => ({
+    display: 'flex'
+});
