+++ src/components/image/styles.ts
@@ -1,0 +1,3 @@
+import { Styles } from '@lapidist/styles';
+
+export const imageStyles = (): Styles => ({});
