+++ src/components/text/index.tsx
@@ -1,21 +1,11 @@
 import React from 'react';
 import deepMerge from 'lodash.merge';
-import { Styles } from '@lapidist/styles';
 import { Box, BoxProps } from '../box';
+import { textStyles } from './styles';
+
+export * from './styles';
 
 export type TextPropType = BoxProps;
-
-export const textStyles = (): Styles => ({
-    fontFamily: 'sans',
-    fontWeight: 'normal',
-    fontSize: '3',
-    lineHeight: 'loose',
-    letterSpacing: 'wide',
-    textColor: {
-        group: 'grey',
-        shade: 'dark'
-    }
-});
 
 export const Text: React.FC<TextPropType> = ({
     as = 'div',
