+++ src/components/text/styles.ts
@@ -1,0 +1,13 @@
+import { Styles } from '@lapidist/styles';
+
+export const textStyles = (): Styles => ({
+    fontFamily: 'sans',
+    fontWeight: 'normal',
+    fontSize: '3',
+    lineHeight: 'loose',
+    letterSpacing: 'wide',
+    textColor: {
+        group: 'grey',
+        shade: 'dark'
+    }
+});
